requirements.txt
flask
flask-sqlalchemy
flask-migrate
mysql-connector-python
python-dotenv